requirements.txt
fastapi
uvicorn
scikit-learn
pandas
numpy
matplotlib
PyYAML
requests
python-multipart
